+++ database-setup.bash
@@ -4,4 +4,5 @@
 psql -d postgres -f ./postgres-setup.sql --echo-all
 
 export NODE_ENV='test'
+npm install
 exit 0
